Can delete tags.

+++ app/src/main/java/com/guillermonegrete/gallery/data/Tag.kt
@@ -5,4 +5,4 @@ import kotlinx.parcelize.Parcelize
 import java.util.*
 
 @Parcelize
-data class Tag(val name: String, val creationDate: Date, val id: Int): Parcelable
+data class Tag(val name: String, val creationDate: Date, val id: Long): Parcelable
